requirements/base.txt
Django==4.0.1
django-environ==0.8.1
djangorestframework==3.13.1
django-cors-headers==3.8.0
djangorestframework-simplejwt==5.0.0
